requirements.txt
pytest==8.3.2
selenium==4.23.1
webdriver-manager==4.0.2
allure-pytest==2.13.5
python-dotenv==1.0.1
